refactor : PieceSymbolMapper 위치 변경
