+++ src/main/java/chess/view/web/PieceSymbolMapper.java
@@ -1,13 +1,11 @@
-package chess.service;
+package chess.view.web;
 
-import chess.domain.board.Board;
 import chess.domain.board.position.Horizontal;
 import chess.domain.board.position.Position;
 import chess.domain.board.position.Vertical;
 import chess.domain.piece.*;
 
 import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Map;
 
 public enum PieceSymbolMapper {
